+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # Change Log
 
+
+### v0.3.0 (2018-08-23)
+- Adds `rb_iterate()` function and fix memory leak
+- Add `roaringbitmap.output_format` parameter
+
 ### v0.2.1 (2018-06-19)
 - Upgrade CRoaring to 0.2.49